dock - move nav components into bark/navigation
