+++ src/app/_components/root-footer.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import { RoutePath } from "@/lib/route-path";
 import Link from "next/link";
 import { useSession } from "next-auth/react";
-import { BarkNavRoute } from "@/components/bark/bark-nav-route";
+import { BarkNavRoute } from "@/components/bark/navigation/bark-nav-route";
 
 const footerRoutes: BarkNavRoute[] = [
   {
