+++ src/components/bark/navigation/bark-dock.tsx
