+++ src/components/bark/navigation/bark-nav-route.ts
